DBZ-8924 Add an abstract LogMiner streaming adapter
